Refactor node filtering logic to ignore unmanaged nodes in Lifecycle and IPAM controllers
- Centralized unmanaged node filtering logic into a shared utility in pkg/util/node/node.go.
- Updated Node Lifecycle Controller to use the shared filtering utility.
- Updated Node IPAM Controller (both Cloud and Range allocators) to ignore nodes with the cloud.google.com/gke-unmanaged-node: true label.
- Updated related tests to reflect the changes.
